chore(no-ticket): CI, strict typing, docs and the Phase 1 decisions
- CI runs the same make targets a developer runs, in the documented order:
  lint -> test-unit -> build-infra -> build-app -> start-infra -> migrate ->
  test-integration -> gen-client drift check -> scan. No CI-only commands.
- mypy strict passes on all production code; tests are exempt from the
  annotation rules only, so the signal there stays the assertion.
- make migrate is idempotent and now also applies the job schema, so the
  worker's tables exist before any worker starts rather than being created
  lazily by whichever one connects first.
- Upgraded Vite and Vitest past their advisories; pip-audit and npm audit
  are both clean.
- CLAUDE.md, and docs updated: plan.md's Phase 1 job platforms corrected to
  the sources decision 6 permits, and technical_boundaries.md section 8 now
  records Clerk as the auth provider with the PaaS and email choices left
  open, plus the five decisions taken while building.

+++ backend/kernel/embeddings/clustering.py
@@ -54,9 +54,9 @@ def cosine_similarity(left: list[float], right: list[float]) -> float:
     """Similarity of two vectors. Pure, so it is unit-testable without numpy."""
     if len(left) != len(right):
         raise ValueError("vectors must have the same length")
-    dot = sum(a * b for a, b in zip(left, right, strict=True))
+    dot = float(sum(a * b for a, b in zip(left, right, strict=True)))
     left_norm = sum(a * a for a in left) ** 0.5
     right_norm = sum(b * b for b in right) ** 0.5
     if left_norm == 0.0 or right_norm == 0.0:
         return 0.0
-    return dot / (left_norm * right_norm)
+    return float(dot / (left_norm * right_norm))
